requirements.txt
cached_property
cliff
coinbase
poloniex
PyYAML
six
stevedore
